- Rename the Authentication files
